+++ ports/templatetags/link_badge.py
@@ -1,23 +1,43 @@
+# Copyright (c) 2020-2026 Danilo G. Baio <dbaio@FreeBSD.org>
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice, this
+# list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+# this list of conditions and the following disclaimer in the documentation
+# and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
 from django import template
-import re
 
 register = template.Library()
 
+BADGES = ['badge-secondary',
+          'badge-info',
+          'badge-light',
+          'badge-dark',
+          'badge-warning',
+          'badge-primary',
+          'badge-success',
+          'badge-danger',
+          ]
+
+
 @register.filter(is_safe=True)
 def get_link_badge(loop_count):
-    badges = ['badge-secondary',
-              'badge-info',
-              'badge-light',
-              'badge-dark',
-              'badge-warning',
-              'badge-primary',
-              'badge-success',
-              'badge-danger',
-              ]
-    while loop_count > len(badges):
-        loop_count = loop_count - len(badges)
+    """Pick a badge class, cycling through the list for long loops"""
 
-    return badges[loop_count - 1]
-
-
-
+    return BADGES[(loop_count - 1) % len(BADGES)]