+++ ports/templatetags/proxy.py
@@ -1,12 +1,47 @@
+# Copyright (c) 2020-2026 Danilo G. Baio <dbaio@FreeBSD.org>
+#
+# Redistribution and use in source and binary forms, with or without
+# modification, are permitted provided that the following conditions are met:
+#
+# 1. Redistributions of source code must retain the above copyright notice, this
+# list of conditions and the following disclaimer.
+#
+# 2. Redistributions in binary form must reproduce the above copyright notice,
+# this list of conditions and the following disclaimer in the documentation
+# and/or other materials provided with the distribution.
+#
+# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+# ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+# WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+# DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE
+# FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+# DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+# SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+# CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
+# OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+# OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
 from django import template
 import re
 
 register = template.Library()
 
+LOG_URL_RE = re.compile(r'http://(.*).nyi.freebsd.org/(.*)', re.IGNORECASE)
+
+
 @register.filter(is_safe=True)
 def get_proxy(url_log):
-    match = re.search('http://(.*).nyi.freebsd.org/(.*)', url_log, re.IGNORECASE)
+    """Rewrite a build server log URL to its pkg-status proxy
+
+    Returns the URL unchanged when it does not point at a nyi.freebsd.org
+    server, which is the case for logs collected from the other clusters.
+    """
+    match = LOG_URL_RE.search(url_log or '')
+    if not match:
+        return url_log
+
     return f'https://pkg-status.freebsd.org/{match.group(1)}/{match.group(2)}'
+
 
 @register.filter(is_safe=True)
 def get_short_name(server):